(3095): Mostly coverity suggested changes
* Did you know I can't spell Sentinel correctly?
* If we're going to use (long double) then MO has to be 16 bytes.
* Many many many many fixes suggested by coverity.  Thanks, coverity!
* These include type problems, sizeof problems, printf format problems,
* Typos, mistakes, memory leaks, buffer overflows, buffer underflows,
* #ifdef mistakes, unreachable code, return value errors,
* Sometimes it was more pedantic than i would care for, but oh well!

+++ include/files.h
@@ -15,7 +15,7 @@
 	int	target_file_write	(const char *, const char *);
 	int	file_writeb 		(int, int, char *);
 	char *	file_read 		(int);
-	char *	file_readb 		(int, int);
+	char *	file_readb 		(int, size_t);
 	int	file_eof 		(int);
 	int	file_close 		(int);
 	int	file_valid		(int);
